+++ demos/dermatology/tests/chartFactory.js
@@ -29,7 +29,7 @@
                     callback(new QuarterPie()
                         .columns(["Category", "Value 1", "Value 2"])
                         .data(Array(20).fill("").map(function (n, i) {
-                            return [String.fromCharCode(65 + i)+" extra label length", Math.random(), Math.random()];
+                            return [String.fromCharCode(65 + i) + " extra label length", Math.random(), Math.random()];
                         }))
                     );
                 });
@@ -109,6 +109,19 @@
                     callback(new Column()
                         .columns(DataFactory.ND.subjects.columns)
                         .data(DataFactory.ND.subjects.data)
+                    );
+                });
+            },
+            domain_percent: function (callback) {
+                legacyRequire(["test/DataFactory", "src/chart/Column"], function (DataFactory, Column) {
+                    callback(new Column()
+                        .columns(["Male/Female Astronauts Per Year", "Female Astronauts", "Male Astronauts"])
+                        .data([["1985", 2, 11], ["1987", 2, 13], ["1990", 5, 18], ["1992", 3, 16], ["1995", 5, 14], ["1996", 8, 27], ["1998", 4, 21], ["2000", 3, 14], ["2004", 2, 9], ["2009", 3, 11]])
+                        .paletteID("FlatUI_Swedish")
+                        .tooltipValueFormat(",.0f")
+                        .showValue(true)
+                        .showValueAsPercent("domain")
+                        .render()
                     );
                 });
             },