+++ demos/gallery/samples/codemirror/XMLEditor.js
@@ -1,0 +1,9 @@
+import { XMLEditor } from "@hpcc-js/codemirror";
+
+const code = `<CATALOG><CD><TITLE>Empire Burlesque</TITLE><ARTIST>Bob Dylan</ARTIST><COUNTRY>USA</COUNTRY><COMPANY>Columbia</COMPANY><PRICE>10.90</PRICE><YEAR>1985</YEAR></CD><CD><TITLE>Hide your heart</TITLE><ARTIST>Bonnie Tyler</ARTIST><COUNTRY>UK</COUNTRY><COMPANY>CBS Records</COMPANY><PRICE>9.90</PRICE><YEAR>1988</YEAR></CD></CATALOG>`;
+
+new XMLEditor()
+    .target("target")
+    .xml(code)
+    .render()
+    ;