+++ packages/graph/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.5.0"></a>
+# [2.5.0](https://github.com/hpcc-systems/Visualization/compare/@hpcc-js/graph@2.4.0...@hpcc-js/graph@2.5.0) (2018-11-26)
+
+
+### Features
+
+* **toolbar:** Add selection group ([d6849ae](https://github.com/hpcc-systems/Visualization/commit/d6849ae))
+
+
+
+
+
+<a name="2.4.0"></a>
+# [2.4.0](https://github.com/hpcc-systems/Visualization/compare/@hpcc-js/graph@2.3.0...@hpcc-js/graph@2.4.0) (2018-11-08)
+
+
+### Features
+
+* **readme:** Add packages/graph/README.md ([a98dfbb](https://github.com/hpcc-systems/Visualization/commit/a98dfbb))
+
+
+
+
+
 <a name="2.3.0"></a>
 # [2.3.0](https://github.com/hpcc-systems/Visualization/compare/@hpcc-js/graph@2.2.1...@hpcc-js/graph@2.3.0) (2018-10-30)